backend/requirements.txt
alembic
boto3
email-validator
fastapi
psycopg[binary]
pydantic
pydantic-settings
python-jose[cryptography]
python-multipart
sqlalchemy
uvicorn[standard]
webauthn>=2.0.0
